pyproject.toml
[project]
name = "icm"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Tsimur Hadeliya", email = "tgadeliya@gmail.com" }
]
requires-python = ">=3.13.3"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=8.4.2",
    "ruff>=0.14.2",
    "ty>=0.0.1a25",
]



[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = [
    "E",
    "F",
    "I",
    "UP"
]
extend-select = [
    "E501",
    "UP045"
]
ignore = [
    "E702"
]
